refactore(folder): structure folder refactoring
