+++ src/components/Loading/LoadingBar.tsx
@@ -1,0 +1,38 @@
+import React, { useState, useEffect } from 'react';
+
+const LoadingBar: React.FC<{ isLoading: boolean }> = ({ isLoading }) => {
+  const [progress, setProgress] = useState(0);
+
+  useEffect(() => {
+    let interval: NodeJS.Timeout;
+    let timeout: NodeJS.Timeout;
+
+    if (isLoading) {
+      setProgress(0);
+      interval = setInterval(() => {
+        setProgress((prev) => (prev < 90 ? prev + 10 : prev));
+      }, 100);
+    } else {
+      setProgress(100);
+      timeout = setTimeout(() => {
+        setProgress(0);
+      }, 500);
+    }
+
+    return () => {
+      clearInterval(interval);
+      clearTimeout(timeout);
+    };
+  }, [isLoading]);
+
+  return (
+    <div className="fixed top-0 left-0 w-full h-1 z-50">
+      <div
+        className="h-full bg-blue-400 opacity-80 transition-all duration-300"
+        style={{ width: `${progress}%` }}
+      ></div>
+    </div>
+  );
+};
+
+export default LoadingBar;
