+++ src/components/Loading/Spinner.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Spinner: React.FC = () => {
+  return (
+    <div className="flex justify-center items-center">
+      <div className="animate-spin rounded-full h-32 w-32 border-t-2 border-b-2 border-blue-500"></div>
+    </div>
+  );
+};
+
+export default Spinner;
