+++ src/components/Login.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '../firebaseConfig';
+import { auth } from '../lib/firebaseConfig';
 import { useNavigate, Link } from 'react-router-dom';
 
 const Login: React.FC = () => {
