+++ src/hooks/useCharacters.ts
@@ -1,0 +1,22 @@
+import useSWR from 'swr';
+
+const fetcher = async (url: string) => {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error('Failed to fetch data');
+  }
+  return response.json();
+};
+
+export const useGhibliPeople = () => {
+  const { data, error } = useSWR('https://ghibliapi.vercel.app/people', fetcher, {
+    revalidateOnFocus: false,
+    refreshInterval: 60000,
+  });
+
+  return {
+    people: data,
+    isLoading: !error && !data,
+    isError: !!error,
+  };
+};
