+++ tailwind.config.js
@@ -6,9 +6,6 @@ export default {
   ],
   theme: {
     extend: {
-      backgroundImage: {
-        'header': "url('assets/images/header.webp')",
-      }
     },
   },
   plugins: [],
